build valve always
